pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vital"
version = "0.1.0"
description = "AI-powered terminal coding assistant using Groq"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "groq>=0.9.0",
    "python-dotenv>=1.0.0",
    "keyring>=24.0.0",
]

[project.scripts]
vital = "vital.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["vital*"]
